pyproject.toml
[tool.poetry]
name = "movielibrary"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12"
fastapi = ">=0.115.14,<0.116.0"
uvicorn = ">=0.35.0,<0.36.0"
sqlalchemy = ">=2.0.41,<3.0.0"
alembic = ">=1.16.2,<2.0.0"
jinja2 = ">=3.1.6,<4.0.0"
python-multipart = ">=0.0.20,<0.0.21"
python-dotenv = ">=1.1.1,<2.0.0"
asyncpg = "^0.30.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
